refactor(backend): wsBridge.ts 60 RPC method + codex broadcast を機能領域別に分離 (#1144 Phase-2)
#1144 Phase-2: wsBridge.ts (2208 行) → 637 行 (-71%) に縮退。
内訳:

## RPC dispatcher 分離 (60 method, 8 ファイル)

- wsHandlers/project.ts (14): loadProject/saveProject/loadScreen/saveScreen/etc.
- wsHandlers/table.ts (8): loadTable/saveTable/listAllTables/loadErLayout/etc.
- wsHandlers/processFlow.ts (22): loadProcessFlow/loadSequence/loadView/
  loadViewDefinition/loadGenericDefinition/loadPageLayout 系
- wsHandlers/misc.ts (8): loadConventions/getFileMtime/getExtensions/
  renameScreenItem/checkScreenItemRefs/etc.
- wsHandlers/workspace.ts (8): workspace.list/status/inspect/open/close/etc.
- wsHandlers/presence.ts (3): presence.heartbeat/list/register
- wsHandlers/editSession.ts (12): editSession.create/save/discard/restoreFromHistory/etc.
- wsHandlers/codex.ts (12): codex.account.*/turn.*/thread.*/model.list/etc.

RpcContext / RpcHandlerMap で統一契約。method → handler の Map lookup に
集約することで dispatcher 自体は 60 LOC まで縮退。

## codex broadcast 分離

- codex/broadcastBridge.ts: CodexBroadcastBridge クラス
  - Lazy CodexConnection singleton (on-demand connect)
  - Notification の WS broadcast
  - Server-initiated request の pending 管理 + resolve/reject

## EditSession service 分離

- wsBridge/editSessionService.ts: EditSessionService クラス
  - editSessionStores / draftHistoryStores の lazy 生成
  - create / attachAsView / detach / setRole / transferEdit / update /
    save / discard / list / fetchPayload / listHistory / restoreFromHistory
  - 1h 周期の cleanup interval + DraftHistory 7 日 TTL

wsBridge.ts 側は MCP tool との契約共有 (#906) のために editSessionXxx
adapter メソッドを保持するが、実装は service に delegate (各 1-3 行)。

## ユーティリティ分離

- wsBridge/killStalePort.ts: killStaleProcessOnPort (Windows + POSIX 両対応)

## 不変

- RPC method 名 (60 件) → frontend store 契約 (変更なし)
- MCP tool 名 → AI client 契約 (変更なし、Phase-1 で既に分離済)
- editSessionXxx 公開 API signature → MCP handlers/editSession.ts 契約 (変更なし)
- broadcast / sendCommand signature (変更なし)
- schemas/ (governance 上 AI 単独変更禁止、当該 PR で触らない)
- 既存 test 513 件 (test code 変更なし、すべて pass)

Refs #1144
